Crate diem_temppath

Structs

TempPath

A simple wrapper for creating a temporary directory that is automatically deleted when it’s dropped. Used in lieu of tempfile due to the large number of dependencies.